Added RecyclerView animation for term assignment tree nodes
